feat(utils): Added configuration file loading tool
- Added file_helper.py module to provide YAML and JSON file loading functions
- Implemented load_yaml_file and load_json_file functions
- Integrated logging and error handling functions

+++ vnpy_tts/gateway/tts_gateway.py
@@ -555,7 +555,7 @@ class TtsTdApi(TdApi):
                 position.yd_volume = data["Position"] - data["TodayPosition"]
 
             # 获取合约的乘数信息
-            size: int = contract.size
+            size: float = contract.size
 
             # 计算之前已有仓位的持仓总成本
             cost: float = position.price * position.volume * size
